This will be the first version released to the beta with notifications fully functioning (I would assume), I fixed the bug that scheduled the period 3/7 notifications at the wrong time. Other bug fixes include:
View not updating when entering foreground.
Crashing when trying to access calendar after changing the month through swiping through days.
Selected day not changing after changing months.
App crashing when unable to connect to Mr. Mandell's server.

Assuming I do not hear about any additional issues, I will submit this to the app store in a couple days. I plan on doing the following in future versions:

Adding a widget.
Changing the way notifications are scheduled; Instead of one view, there will be a menu that will allow you to schedule notifications for flex/lunch and before the end of class instead of before the beginning. I also might add a place to input room numbers so that you can be reminded, I got a few requests for it but I don't really see the point after the first month or so of school, thoughts?

contact me at stephen@nwlewis.com