UNINSTALL AND REINSTALL. I changed how I stored information, so to get rid of any old information you should uninstall the app.
I added a widget, if you want to use/test it, swipe to the left until you see your widgets. To add the widget, scroll to the bottom till you see the edit button. Press it and add the Bell Schedule widget.

I will say again that this is hopefully the last beta before submitting it for review again.