UNINSTALL AND REINSTALL. I changed how I stored information, so to get rid of any old information, you should uninstall the app.
I added a widget, if you want to use/test it, swipe to the left until you see your widgets. To add the widget, scroll to the bottom till you see the edit button. Press it and add the Bell Schedule widget.

I intend to have this submitted for review (so it can be put on the app store) at the same time this comes out, so if you find any bugs it would be helpful to know asap so I can un-submit and fix it. Once again my e-mail is stephen@nwlewis.com